+++ spec/ig_markets/dealing_platform_spec.rb
@@ -28,7 +28,7 @@ describe IGMarkets::DealingPlatform do
     deal_confirmation = build :deal_confirmation
 
     expect(session).to receive(:get).with('confirms/deal_id', IGMarkets::API_V1).and_return(deal_confirmation)
-    expect(platform.deal_confirmation('deal_id')).to eq(deal_confirmation)
+    expect(platform.deal_confirmation(deal_confirmation.deal_id)).to eq(deal_confirmation)
   end
 
   it 'can retrieve the current applications' do